TEL-10545 Voipmonitor not restarted because of pid file deletion.
Summary:
This is the second part of the solution. It assures that pid file is created with the proper permissions.

Solution:
Voipmonitor has been modified for changing permissions when the pid file is created. Note that this solution will not modify the script provided on the last modification but complement it. Stop procedure in script needs to be preserved to assure successful pushing.

Test Plan: Solution was successfully tested by confirming permissions properly set on a system with previous invalid settings, and all usual cases for stopping, starting, restarting service with and without monit.

Reviewers: corey, sega

Subscribers: TelPlat

Differential Revision: https://phabricator.fspot.us/D13423