TEL-13668 - Removing the automatic printing of stacktrace that's hanging voipmon
Summary:
Today, we're seeing plenty of alerts about voipmon not working.
This is due to it receiving a signal and attempting to print a backtrace manually.
This change removes the functionality, instead relaying on the coredump and debug symbols in the runtime binary to acheive the same.

Test Plan: Make sure it's still working after (duh).

Reviewers: nestor, abbasali, mghoneim, max

Reviewed By: nestor

Subscribers: TelPlat

JIRA Issues: TEL-13668

Differential Revision: https://phabricator.fspot.us/D15832